Upate wdt

+++ lib/drivers/include/wdt.h
@@ -145,7 +145,7 @@ void wdt_start(wdt_device_number_t id, uint64_t time_out_ms, plic_irq_callback_t
  * @return      Wdt time
  *
  */
-uint64_t wdt_init(wdt_device_number_t id, uint64_t time_out_ms, plic_irq_callback_t on_irq, void *ctx);
+uint32_t wdt_init(wdt_device_number_t id, uint64_t time_out_ms, plic_irq_callback_t on_irq, void *ctx);
 
 /**
  * @brief       Stop wdt
